pyproject.toml
[tool.poetry]
name = "configly"
version = "0.1.0"
description = ""
authors = []
license = "Apache-2.0"
keywords = [ "config", "yaml", "toml", "env" ]
repository = "https://github.com/schireson/configly"
packages = [
    { from = "src", include = "configly" },
]
readme = 'README.md'
include = [
    "*.md",
]

[tool.poetry.dependencies]
python = ">=3,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*,!=3.4.*"

'ruamel.yaml' = { version = "*", optional = true }
toml = { version = "*", optional = true }
hvac = { version = "*", optional = true }

[tool.poetry.dev-dependencies]
black = { version = "19.3b0", allows-prereleases = true, python = ">=3.6" }
flake8 = "^3.7"
isort = "^4.3"
pytest = "^4.4"
pytest-cov = "^2.6"
responses = "*"

[tool.poetry.extras]
yaml = [ "ruamel.yaml" ]
toml = [ "toml" ]
vault = [ "hvac" ]

[tool.black]
line_length = 100
target_version = [ 'py34', 'py35', 'py36', 'py37' ]

[tool.isort]
default_section = 'FIRSTPARTY'
include_trailing_comma = true
indent = '    '
length_sort = false
line_length = 88
multi_line_output = 3
order_by_type = false
known_first_party = 'tests'
use_parentheses = true

[build-system]
requires = [ "poetry>=0.12" ]
build-backend = "poetry.masonry.api"
